Adjust GUI
The existing GUI is copied straight from the JavaFX tutorial.

The existing GUI is too basic and not nice to look at.

Let's change the background color, the shape of the avatar and the font used.

It needs to be changed into more aesthestically pleasing GUI and to also add personality to the chatbot.